refactor(frontend): extract the account form logic
AccountForm.vue held 331 script lines doing two separate jobs behind one
template, which is what made it the file the testing guide names as the
cost of testing a fat component first.

Splits them into use-account-form-warnings (which services and chains
are worth warning about before an account is added, and how many of
those warnings are shown at once) and use-account-form-state (the
transitions that replace the whole AccountManageState, since the chain
implies the shape of data). The component keeps the template, the
validate delegation and the wiring, dropping from 430 to 161 lines.

No behaviour change: the existing spec passes untouched.
